test: add test-utils package and move fake-did-provider there
