+++ __browser_tests__/react-sample/package.json
@@ -6,19 +6,19 @@
     "@types/node": "16.11.36",
     "@types/react": "18.0.9",
     "@types/react-dom": "18.0.5",
-    "@veramo/core": "^3.1.4",
-    "@veramo/credential-w3c": "^3.1.4",
-    "@veramo/data-store": "^3.1.4",
-    "@veramo/did-comm": "^3.1.4",
-    "@veramo/did-jwt": "^3.1.4",
-    "@veramo/did-manager": "^3.1.4",
-    "@veramo/did-provider-ethr": "^3.1.4",
-    "@veramo/did-provider-key": "^3.1.4",
-    "@veramo/did-provider-web": "^3.1.4",
-    "@veramo/did-resolver": "^3.1.4",
-    "@veramo/key-manager": "^3.1.4",
-    "@veramo/message-handler": "^3.1.4",
-    "@veramo/selective-disclosure": "^3.1.4",
+    "@veramo/core": "^3.1.0",
+    "@veramo/credential-w3c": "^3.1.0",
+    "@veramo/data-store": "^3.1.0",
+    "@veramo/did-comm": "^3.1.0",
+    "@veramo/did-jwt": "^3.1.0",
+    "@veramo/did-manager": "^3.1.0",
+    "@veramo/did-provider-ethr": "^3.1.0",
+    "@veramo/did-provider-key": "^3.1.0",
+    "@veramo/did-provider-web": "^3.1.0",
+    "@veramo/did-resolver": "^3.1.0",
+    "@veramo/key-manager": "^3.1.0",
+    "@veramo/message-handler": "^3.1.0",
+    "@veramo/selective-disclosure": "^3.1.0",
     "crypto": "npm:crypto-browserify",
     "did-resolver": "3.2.0",
     "ethr-did-resolver": "5.0.4",
