+++ packages/data-store/src/entities/PreMigrationEntities.ts
@@ -1,4 +1,5 @@
 import { BaseEntity, Column, Entity, PrimaryColumn } from 'typeorm'
+import { KeyType } from "./key";
 
 @Entity('key')
 export class PreMigrationKey extends BaseEntity {
