+++ packages/test-utils/README.md
@@ -1,0 +1,3 @@
+# Veramo test-utils
+
+Helper methods for Veramo integration tests. This package is NOT meant to be used in production.
