+++ packages/test-utils/package.json
@@ -1,0 +1,35 @@
+{
+  "name": "@veramo/test-utils",
+  "private": true,
+  "description": "Helper methods for Veramo integration tests. This package is not meant to be used in production.",
+  "version": "3.1.4",
+  "main": "build/index.js",
+  "types": "build/index.d.ts",
+  "scripts": {
+    "build": "tsc"
+  },
+  "dependencies": {
+    "@veramo/core": "^3.1.0",
+    "@veramo/data-store-json": "^3.1.0",
+    "@veramo/did-manager": "^3.1.0",
+    "@veramo/key-manager": "^3.1.0",
+    "@veramo/utils": "^3.1.0",
+    "did-resolver": "3.2.0"
+  },
+  "devDependencies": {
+    "typescript": "4.7.2"
+  },
+  "files": [
+    "build/**/*",
+    "src/**/*",
+    "README.md",
+    "LICENSE"
+  ],
+  "publishConfig": {
+    "access": "public"
+  },
+  "repository": "git@github.com:uport-project/veramo.git",
+  "author": "Mircea Nistor <mircea.nistor@mesh.xyz>",
+  "license": "Apache-2.0",
+  "keywords": []
+}
