+++ packages/test-utils/src/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Contains helper methods for Veramo integration tests.
+ *
+ * NOT to be used in production.
+ *
+ * @packageDocumentation
+ */
+
+export * from './fake-did'
