+++ packages/test-utils/tsconfig.json
@@ -1,0 +1,14 @@
+{
+  "extends": "../tsconfig.settings.json",
+  "compilerOptions": {
+    "rootDir": "src",
+    "outDir": "build",
+    "declarationDir": "build"
+  },
+  "references": [
+    { "path": "../core" },
+    { "path": "../data-store-json" },
+    { "path": "../did-manager" },
+    { "path": "../key-manager" },
+  ]
+}
