+++ packages/tsconfig.settings.json
@@ -4,6 +4,7 @@
     "preserveConstEnums": true,
     "sourceMap": true,
     "target": "es6",
+    "lib": ["ES2019"],
     "module": "commonjs",
     "moduleResolution": "node",
     "esModuleInterop": true,
